@bjorn3 2025-09-10
Have you benchmarked this on LEB128 integers obtained from actual debuginfo? They almost certainly have a different distribution to what rustc sees. And how does this compare outside of micro-benchmarks (eg reading debuginfo for an entire executable)?