@d-e-s-o 2025-09-15
> I tried running the existing benchmark on my system, and this actually makes the most relevant one slower, which is surprising

That is surprising indeed. I am not sure where that comes from. I compared at the assembly but I am none the wiser, really. The only explanation I can muster would be that, because the compiler seems to generate more code (not sure it if unrolls more iterations or so), we may end up putting more pressure on the L1i cache, perhaps exaggerated by inlining. That could explain why microbenchmarks consistently show improvements, I'd think. In any event, performance counters aren't really conclusive, as numbers seem to be all over the place and not really correlated with reported wall clock numbers to the degree I can tell.

I can't spend much more time on the investigation here, so I kept the main loop body as-is now and just moved the first check out of the loop. That does seem to improve performance, but it's not much anymore. Take it or leave it.

> If we're going to change unsigned, then we should change signed too.

Why can't we optimize one without the other...?