@philipc 2025-09-16
In the commit message for 3c56f86, are the After and Before labels wrong?

This change looks good to me now. I see a performance improvement on my system for `bench_parsing_debug_info`, and the `addr2line` benchmarks also show an improvement:

```
context_new             time:   [488.19 µs 488.65 µs 489.15 µs]
                        change: [−2.5561% −1.7932% −1.1502%] (p = 0.00 < 0.05)
                        Performance has improved.

context_new_parse_lines time:   [5.1523 ms 5.1623 ms 5.1731 ms]
                        change: [−1.1673% −0.9021% −0.6303%] (p = 0.00 < 0.05)
                        Change within noise threshold.

context_new_parse_functions
                        time:   [7.5898 ms 7.5999 ms 7.6105 ms]
                        change: [−4.9737% −4.7767% −4.5780%] (p = 0.00 < 0.05)
                        Performance has improved.

context_new_parse_inlined_functions
                        time:   [26.468 ms 26.505 ms 26.544 ms]
                        change: [−2.5977% −2.2868% −2.0112%] (p = 0.00 < 0.05)
                        Performance has improved.
```

> Why can't we optimize one without the other...?

We can, but I meant that if it is worth doing for one then it is likely to be worth doing for the other. However looking into it, the sign extension does complicate things, so it's not so clear.

For reference, https://github.com/rust-lang/rust/pull/92604 is now the most relevant rust PR here, and it didn't do a similar change for signed, but makes no mention of considering it.