@d-e-s-o 2025-09-16
> In the commit message for [3c56f86](https://github.com/gimli-rs/gimli/commit/3c56f868738bf3f7cbb4fbf87737878af093998c), are the After and Before labels wrong?

Ehm, yes. I fixed that.

> This change looks good to me now. I see a performance improvement on my system for `bench_parsing_debug_info`, and the `addr2line` benchmarks also show an improvement:

:+1: 

> > Why can't we optimize one without the other...?
> 
> We can, but I meant that if it is worth doing for one then it is likely to be worth doing for the other. However looking into it, the sign extension does complicate things, so it's not so clear.

Yeah, that was also the sticking point for me when I looked at it.

If there are no further comments, feel free to merge.